Fixed issue with missing import files
