+++ Assets/Scripts/Scriptable Objects/Behaviours.cs
@@ -6,25 +6,24 @@ using UnityEngine.AI;
 public abstract class AIBehaviour : ScriptableObject
 {
 	public abstract void Act(ref Knowledge knowledge);
+	public abstract void Entry(ref Knowledge knowledge);
+	public abstract void Exit(ref Knowledge knowledge);
 }
 
 public class Knowledge
 {
-	public Vector3? dangerPosition;
-
 	public Transform AITransform { get; private set; }
 	public Vector3 PlayerPosition { get; private set; }
-	public FurnitureSO Stats { get; private set; }
+	public FurnitureSO Info { get; private set; }
 	public NavMeshAgent Agent { get; private set; }
 	public SoundAlert? MostProminentSound { get; private set; }
 	public bool CanSeePlayer { get; private set; }
 
-	public Knowledge(Transform t, Vector3 p, Vector3? d, FurnitureSO s, NavMeshAgent a, SoundAlert? sa, bool v)
+	public Knowledge(Transform t, Vector3 p, FurnitureSO i, NavMeshAgent a, SoundAlert? sa, bool v)
 	{
 		AITransform = t;
 		PlayerPosition = p;
-		dangerPosition = d;
-		Stats = s;
+		Info = i;
 		Agent = a;
 		MostProminentSound = sa;
 		CanSeePlayer = v;
